pyproject.toml
[project]
name = "photoorganizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.0.0",
    "pyinstaller>=6.16.0",
    "pyqt5==5.15.9",
    "pyqt5-qt5==5.15.2",
]
